FREEPBX-13469 errors when no sound packages are installed or they are missing

+++ assets/js/recordings.js
@@ -20,6 +20,9 @@ $("#recordings-frm").submit(function(e) {
 	if($("#name").val().trim() === "") {
 		return warnInvalid($("#name"),_("You must set a valid name for this recording"));
 	}
+	if(!langs.length) {
+		return warnInvalid($("#language"),_("No languages are defined! You must define at least one in the Sound Languages module!"));
+	}
 	data.name = $("#name").val().trim();
 
 	data.id = $("#id").val();
@@ -830,8 +833,13 @@ function addFile(name, filenames, languages, temp, exists) {
  * @return {string}       Return the resulting html for the cel
  */
 function linkFormatter(value, row, index){
-	var html = '<a href="?display=recordings&action=edit&id='+row.id+'"><i class="fa fa-pencil"></i></a>';
-	html += '&nbsp;<a href="?display=recordings&action=delete&id='+row.id+'" class="delAction"><i class="fa fa-trash"></i></a>';
+	var html = '';
+	if(langs.length > 0) {
+		html += '<a href="?display=recordings&action=edit&id='+row.id+'"><i class="fa fa-pencil"></i></a>';
+		html += '&nbsp;<a href="?display=recordings&action=delete&id='+row.id+'" class="delAction"><i class="fa fa-trash"></i></a>';
+	} else {
+		html += _('Please add at least one Sound Package in the Sound Languages Module');
+	}
 	return html;
 }
